Limit the size of proposals.
The Juno mainnet allows storage writes in execute messages that are
larger than the amount of data queriable in a query message. This
means that a sufficently large proposal (around 40,000 bytes) will
unqueriable after being created.

To prevent this this commit limits the size, in bytes, that a proposal
may be. A sensible limit for this size was found by repeatedly storing
a contract on the mainnet and creating the largest possible proposal.