Add min_voting_period to cw-proposal-single.
This adds a minimum voting period. Having a minimum voting period
allows DAO members some time to respond in the event of a DAO hostile
takeover.

For example, say a DAO exists with some tokens in a LP pool. An
attacker may purchase a large number of those tokens and stake them to
take over the DAO and give the DAO's treasury to themselves. If a
minimum voting period is configured DAO members will have some time to
see this occuring and remove liquidity from the pool to reduce the
amount of damage the attacker might do.