Make admin non-optional and default admin to core contract.
Where we used to set the admin to `None` we now set the admin to the
address of the contract itself. This reduces the chance that a typo
prevents the DAO from setting an admin and allows setting the admin
again after an admin has been removed (for example, if a subDAO
becomes its own DAO and then joins another parent DAO).